Making the data_loader fn to return an arbitrary #loaders was wrong
(1) Since, if we needed only one loader (e.g., test_set for final evaluation), then it would not return a loader but a single item list/tuple/GEN/etc.
(2) It's better to use get_loaders fn only when we want two or more loaders (e.g., train & val for train).
(3) When we want only one loader then it would be better to use torch.data.DataLoader directly (withou using a fn).
This is OK as we should not touch our test set before we're done with training and hyperparameter-finetuning. So only need it for evaluation/inference.