multiple changes to image_mask_plot, plot_metrics, and plot_metrics_finetune
(1) the plot_metrics and plot_metrics_finetune now allow to change the scale of the y-axis, cropping y-axis, and setting an interval for x-axis.
(2) the plot_metrics and plot_metrics_finetune now don't need the epochs arg, as they calculate it inside.
(3) the image_mask_plot can now shuffle the input tensors to show a randomize set of images at consecutive runs.